refactor context and reducers
