+++ app/javascript/layouts/ApplicationTimeout.ts
@@ -2,8 +2,7 @@ import { createElement, useContext } from "react"
 
 import { IdleTimeout, lRoute, t } from "@bloom-housing/ui-components"
 
-// import { AppSubmissionContext } from "../../../lib/AppSubmissionContext"
-import { UserContext } from "../authentication/UserContext"
+import UserContext from "../authentication/context/UserContext"
 
 const onTimeout = () => {
   // conductor.reset()
@@ -11,22 +10,20 @@ const onTimeout = () => {
 
 const ApplicationTimeout = () => {
   const { profile } = useContext(UserContext)
-  // const { conductor } = useContext(AppSubmissionContext)
+
+  if (!profile) return null
 
   // Only return something if the user is not logged in - otherwise we'll let the root logged in user timeout handle
   // things.
-  return (
-    profile &&
-    createElement(IdleTimeout, {
-      promptTitle: t("t.areYouStillWorking"),
-      promptText: t("application.timeout.text"),
-      promptAction: t("application.timeout.action"),
-      redirectPath: lRoute("/"),
-      alertMessage: t("application.timeout.afterMessage"),
-      alertType: "alert",
-      onTimeout,
-    })
-  )
+  return createElement(IdleTimeout, {
+    promptTitle: t("t.areYouStillWorking"),
+    promptText: t("application.timeout.text"),
+    promptAction: t("application.timeout.action"),
+    redirectPath: lRoute("/"),
+    alertMessage: t("application.timeout.afterMessage"),
+    alertType: "alert",
+    onTimeout,
+  })
 }
 
-export { ApplicationTimeout as default, ApplicationTimeout }
+export default ApplicationTimeout
