+++ app/javascript/layouts/FormLayout.tsx
@@ -1,25 +1,23 @@
 import React from "react"
 
-import { ApplicationTimeout } from "./ApplicationTimeout"
+import ApplicationTimeout from "./ApplicationTimeout"
 import Layout from "./Layout"
 
 export interface FormLayoutProps {
   children: React.ReactNode
 }
 
-const FormLayout = (props: FormLayoutProps) => {
-  return (
-    <>
-      <ApplicationTimeout />
-      <Layout>
-        <section className="bg-gray-300">
-          <div className="md:mb-20 md:mt-12 mx-auto max-w-lg print:my-0 print:max-w-full">
-            {props.children}
-          </div>
-        </section>
-      </Layout>
-    </>
-  )
-}
+const FormLayout = (props: FormLayoutProps) => (
+  <>
+    <ApplicationTimeout />
+    <Layout>
+      <section className="bg-gray-300">
+        <div className="md:mb-20 md:mt-12 mx-auto max-w-lg print:my-0 print:max-w-full">
+          {props.children}
+        </div>
+      </section>
+    </Layout>
+  </>
+)
 
 export default FormLayout
