+++ app/javascript/pages/SignIn.tsx
@@ -1,17 +1,12 @@
 import React from "react"
 
 import { SignInForm } from "../authentication/SignInForm"
-import { LoggedInUserIdleTimeout } from "../authentication/timeout"
-import { UserProvider } from "../authentication/UserContext"
 import FormsLayout from "../layouts/FormLayout"
 import withAppSetup from "../layouts/withAppSetup"
 
 const SignIn = () => (
   <FormsLayout>
-    <UserProvider>
-      <LoggedInUserIdleTimeout onTimeout={() => console.log("Logout")} />
-      <SignInForm />
-    </UserProvider>
+    <SignInForm />
   </FormsLayout>
 )
